+++ example/src/ack-angular/components.ts
@@ -7,7 +7,7 @@ import {
   ElementRef
 } from "@angular/core"
 
-import { AckModal } from "./AckModal.component"
+import { AckModal } from "./components/AckModal.component"
 import { AckArray } from "./AckArray.directive"
 import { AckOptions } from "./AckOptions.component"
 import { AckOptionsModal } from "./AckOptionsModal.component"